# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Document Intelligence Pipeline API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type JobStatus = (typeof JobStatus)[keyof typeof JobStatus];

export const JobStatus = {
  pending: "pending",
  processing: "processing",
  completed: "completed",
  failed: "failed",
} as const;

export interface CreateJobRequest {
  name: string;
  description?: string;
}

export interface Job {
  id: string;
  name: string;
  description?: string | null;
  status: JobStatus;
  fileCount: number;
  createdAt: string;
  updatedAt: string;
  errorMessage?: string | null;
}

export interface UploadedFile {
  id: string;
  jobId: string;
  filename: string;
  mimeType: string;
  sizeBytes: number;
  createdAt: string;
}

export interface TimelineEvent {
  id: string;
  date?: string | null;
  dateInferred: boolean;
  dateConfidence?: string | null;
  description: string;
  sourceFile: string;
  sourceLocation?: string | null;
  entities: string[];
  eventType?: string | null;
}

export type RagChunkMetadata = {
  date?: string | null;
  source: string;
  entities: string[];
  eventType?: string | null;
};

export interface RagChunk {
  id: string;
  text: string;
  metadata: RagChunkMetadata;
}

export interface FineTuneExample {
  input: string;
  output: string;
  context: string;
}

export type JobOutputsValidationReport = {
  totalFiles: number;
  processedFiles: number;
  totalEvents: number;
  totalEntities: number;
  conflictsDetected: number;
  dataLossRisk: string;
};

export interface JobOutputs {
  /** Full chronological master document text */
  masterDocument: string;
  ragChunks: RagChunk[];
  fineTuneExamples: FineTuneExample[];
  validationReport: JobOutputsValidationReport;
}

export type JobDetail = Job & {
  files: UploadedFile[];
};

export type ListJobs200 = {
  jobs: Job[];
};

export type GetJobTimeline200 = {
  events: TimelineEvent[];
};

export type UploadFileBody = {
  file: Blob;
};