# lib/api-zod/src/generated/types/createJobRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Document Intelligence Pipeline API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateJobRequest {
  name: string;
  description?: string;
}