# lib/api-zod/src/generated/types/fineTuneExample.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Document Intelligence Pipeline API
 * OpenAPI spec version: 0.1.0
 */

export interface FineTuneExample {
  input: string;
  output: string;
  context: string;
}