# lib/api-zod/src/generated/types/job.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Document Intelligence Pipeline API
 * OpenAPI spec version: 0.1.0
 */
import type { JobStatus } from "./jobStatus";

export interface Job {
  id: string;
  name: string;
  description?: string | null;
  status: JobStatus;
  fileCount: number;
  createdAt: Date;
  updatedAt: Date;
  errorMessage?: string | null;
}