# lib/api-zod/src/generated/types/jobDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Document Intelligence Pipeline API
 * OpenAPI spec version: 0.1.0
 */
import type { Job } from "./job";
import type { UploadedFile } from "./uploadedFile";

export type JobDetail = Job & {
  files: UploadedFile[];
};