# lib/api-zod/src/generated/types/jobOutputs.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Document Intelligence Pipeline API
 * OpenAPI spec version: 0.1.0
 */
import type { FineTuneExample } from "./fineTuneExample";
import type { JobOutputsValidationReport } from "./jobOutputsValidationReport";
import type { RagChunk } from "./ragChunk";

export interface JobOutputs {
  /** Full chronological master document text */
  masterDocument: string;
  ragChunks: RagChunk[];
  fineTuneExamples: FineTuneExample[];
  validationReport: JobOutputsValidationReport;
}