# lib/api-zod/src/generated/types/jobStatus.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Document Intelligence Pipeline API
 * OpenAPI spec version: 0.1.0
 */

export type JobStatus = (typeof JobStatus)[keyof typeof JobStatus];

export const JobStatus = {
  pending: "pending",
  processing: "processing",
  completed: "completed",
  failed: "failed",
} as const;