# lib/api-zod/src/generated/types/ragChunk.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Document Intelligence Pipeline API
 * OpenAPI spec version: 0.1.0
 */
import type { RagChunkMetadata } from "./ragChunkMetadata";

export interface RagChunk {
  id: string;
  text: string;
  metadata: RagChunkMetadata;
}