# lib/api-zod/src/generated/types/timelineEvent.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Document Intelligence Pipeline API
 * OpenAPI spec version: 0.1.0
 */

export interface TimelineEvent {
  id: string;
  date?: string | null;
  dateInferred: boolean;
  dateConfidence?: string | null;
  description: string;
  sourceFile: string;
  sourceLocation?: string | null;
  entities: string[];
  eventType?: string | null;
}