# lib/api-zod/src/generated/types/uploadedFile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Document Intelligence Pipeline API
 * OpenAPI spec version: 0.1.0
 */

export interface UploadedFile {
  id: string;
  jobId: string;
  filename: string;
  mimeType: string;
  sizeBytes: number;
  createdAt: Date;
}